pyaiot/broker: reorganize files
